Branched + simplified repo for paper reproduction
