--- byol/dataloading/format_test.py
@@ -1,7 +1,0 @@
-import torch
-
-dict = {
-    "a": 1,
-    "b": 2,
-    "c": 3,
-}
