+++ byol/resnet.py
@@ -1,0 +1,322 @@
+from typing import Type, Any, Callable, Union, List, Optional
+import torch
+import torch.nn as nn
+from torch import Tensor
+
+# Largely lifted from torchvision.models
+# https://pytorch.org/vision/stable/_modules/torchvision/models/resnet.html#resnet18
+
+
+def conv3x3(
+    in_planes: int, out_planes: int, stride: int = 1, groups: int = 1, dilation: int = 1
+) -> nn.Conv2d:
+    """3x3 convolution with padding"""
+    return nn.Conv2d(
+        in_planes,
+        out_planes,
+        kernel_size=3,
+        stride=stride,
+        padding=dilation,
+        groups=groups,
+        bias=False,
+        dilation=dilation,
+    )
+
+
+def conv1x1(in_planes: int, out_planes: int, stride: int = 1) -> nn.Conv2d:
+    """1x1 convolution"""
+    return nn.Conv2d(in_planes, out_planes, kernel_size=1, stride=stride, bias=False)
+
+
+class BasicBlock(nn.Module):
+    expansion: int = 1
+
+    def __init__(
+        self,
+        inplanes: int,
+        planes: int,
+        stride: int = 1,
+        downsample: Optional[nn.Module] = None,
+        groups: int = 1,
+        base_width: int = 64,
+        dilation: int = 1,
+        norm_layer: Optional[Callable[..., nn.Module]] = None,
+    ) -> None:
+        super().__init__()
+        if norm_layer is None:
+            norm_layer = nn.BatchNorm2d
+        if groups != 1 or base_width != 64:
+            raise ValueError("BasicBlock only supports groups=1 and base_width=64")
+        if dilation > 1:
+            raise NotImplementedError("Dilation > 1 not supported in BasicBlock")
+        # Both self.conv1 and self.downsample layers downsample the input when stride != 1
+        self.conv1 = conv3x3(inplanes, planes, stride)
+        self.bn1 = norm_layer(planes)
+        self.relu = nn.ReLU(inplace=True)
+        self.conv2 = conv3x3(planes, planes)
+        self.bn2 = norm_layer(planes)
+        self.downsample = downsample
+        self.stride = stride
+
+    def forward(self, x: Tensor) -> Tensor:
+        identity = x
+
+        out = self.conv1(x)
+        out = self.bn1(out)
+        out = self.relu(out)
+
+        out = self.conv2(out)
+        out = self.bn2(out)
+
+        if self.downsample is not None:
+            identity = self.downsample(x)
+
+        out += identity
+        out = self.relu(out)
+
+        return out
+
+
+# Something is up when using Bottleneck layers - need to debug this.
+class Bottleneck(nn.Module):
+    # Bottleneck in torchvision places the stride for downsampling at 3x3 convolution(self.conv2)
+    # while original implementation places the stride at the first 1x1 convolution(self.conv1)
+    # according to "Deep residual learning for image recognition"https://arxiv.org/abs/1512.03385.
+    # This variant is also known as ResNet V1.5 and improves accuracy according to
+    # https://ngc.nvidia.com/catalog/model-scripts/nvidia:resnet_50_v1_5_for_pytorch.
+
+    expansion: int = 4
+
+    def __init__(
+        self,
+        inplanes: int,
+        planes: int,
+        stride: int = 1,
+        downsample: Optional[nn.Module] = None,
+        groups: int = 1,
+        base_width: int = 64,
+        dilation: int = 1,
+        norm_layer: Optional[Callable[..., nn.Module]] = None,
+    ) -> None:
+        super().__init__()
+        if norm_layer is None:
+            norm_layer = nn.BatchNorm2d
+        width = int(planes * (base_width / 64.0)) * groups
+        # Both self.conv2 and self.downsample layers downsample the input when stride != 1
+        self.conv1 = conv1x1(inplanes, width)
+        self.bn1 = norm_layer(width)
+        self.conv2 = conv3x3(width, width, stride, groups, dilation)
+        self.bn2 = norm_layer(width)
+        self.conv3 = conv1x1(width, planes * self.expansion)
+        self.bn3 = norm_layer(planes * self.expansion)
+        self.relu = nn.ReLU(inplace=True)
+        self.downsample = downsample
+        self.stride = stride
+
+    def forward(self, x: Tensor) -> Tensor:
+        identity = x
+
+        out = self.conv1(x)
+        out = self.bn1(out)
+        out = self.relu(out)
+
+        out = self.conv2(out)
+        out = self.bn2(out)
+        out = self.relu(out)
+
+        out = self.conv3(out)
+        out = self.bn3(out)
+
+        if self.downsample is not None:
+            identity = self.downsample(x)
+
+        out += identity
+        out = self.relu(out)
+
+        return out
+
+
+class ResNet(nn.Module):
+    def __init__(
+        self,
+        block: Type[Union[BasicBlock, Bottleneck]],
+        layers: List[int],
+        n_c: int = 3,
+        downscale: bool = False,
+        features: int = 512,
+        zero_init_residual: bool = False,
+        groups: int = 1,
+        width_per_group: int = 64,
+    ) -> None:
+        super().__init__()
+        self.norm_layer = nn.BatchNorm2d
+
+        self.layers = nn.ModuleList([])
+
+        self.inplanes = 64
+        self.dilation = 1
+        self.groups = groups
+        self.base_width = width_per_group
+        self.features = features
+        self.dim = features
+
+        # Change to smaller kernel for first layer for lower resolution data (e.g. CIFAR10)
+        if downscale:
+            conv1 = nn.Conv2d(n_c, self.inplanes, kernel_size=3, stride=1, padding=1, bias=False)
+        else:
+            conv1 = nn.Conv2d(n_c, self.inplanes, kernel_size=7, stride=2, padding=3, bias=False)
+
+        self.layers.append(
+            nn.Sequential(
+                conv1,
+                self.norm_layer(self.inplanes),
+                nn.ReLU(inplace=True),
+                nn.MaxPool2d(kernel_size=3, stride=2, padding=1),
+            )
+        )
+
+        # self.layers.append(self._make_layer(block, 64, layers[0]))
+        # self.layers.append(self._make_layer(block, 128, layers[1], stride=2))
+        # self.layers.append(self._make_layer(block, 256, layers[2], stride=2))
+        # self.layers.append(self._make_layer(block, 512, layers[3], stride=2))
+
+        # Add resnet layers
+        self.layers += self._make_layer(block, 64, layers[0])
+        self.layers += self._make_layer(block, 128, layers[1], stride=2)
+        self.layers += self._make_layer(block, 256, layers[2], stride=2)
+        self.layers += self._make_layer(block, 512, layers[3], stride=2)
+
+        if features != 512:
+            # Conv layer simply projects features down to stated dimension with 1x1 kernel
+            self.layers[-1] = block(
+                512,
+                features,
+                groups=self.groups,
+                base_width=self.base_width,
+                dilation=self.dilation,
+                norm_layer=self.norm_layer,
+            )
+
+        self.layers.append(nn.AdaptiveAvgPool2d((1, 1)))
+
+        # Point finetuner to module list of layers
+        self.finetuning_layers = self.layers[:-1]
+
+        for m in self.modules():
+            if isinstance(m, nn.Conv2d):
+                nn.init.kaiming_normal_(m.weight, mode="fan_out", nonlinearity="relu")
+            elif isinstance(m, (nn.BatchNorm2d, nn.GroupNorm)):
+                nn.init.constant_(m.weight, 1)
+                nn.init.constant_(m.bias, 0)
+
+        # Zero-initialize the last BN in each residual branch,
+        # so that the residual branch starts with zeros, and each residual block behaves like an identity.
+        # This improves the model by 0.2~0.3% according to https://arxiv.org/abs/1706.02677
+        if zero_init_residual:
+            for m in self.modules():
+                if isinstance(m, Bottleneck) and m.bn3.weight is not None:
+                    nn.init.constant_(m.bn3.weight, 0)  # type: ignore[arg-type]
+                elif isinstance(m, BasicBlock) and m.bn2.weight is not None:
+                    nn.init.constant_(m.bn2.weight, 0)  # type: ignore[arg-type]
+
+    def _make_layer(
+        self,
+        block: Type[Union[BasicBlock, Bottleneck]],
+        planes: int,
+        blocks: int,
+        stride: int = 1,
+        dilate: bool = False,
+    ) -> nn.ModuleList:
+        norm_layer = self.norm_layer
+        downsample = None
+        previous_dilation = self.dilation
+        if dilate:
+            self.dilation *= stride
+            stride = 1
+        if stride != 1 or self.inplanes != planes * block.expansion:
+            downsample = nn.Sequential(
+                conv1x1(self.inplanes, planes * block.expansion, stride),
+                norm_layer(planes * block.expansion),
+            )
+
+        layers = []
+        layers.append(
+            block(
+                self.inplanes,
+                planes,
+                stride,
+                downsample,
+                self.groups,
+                self.base_width,
+                previous_dilation,
+                norm_layer,
+            )
+        )
+        self.inplanes = planes * block.expansion
+        for _ in range(1, blocks):
+            layers.append(
+                block(
+                    self.inplanes,
+                    planes,
+                    groups=self.groups,
+                    base_width=self.base_width,
+                    dilation=self.dilation,
+                    norm_layer=norm_layer,
+                )
+            )
+
+        # return nn.Sequential(*layers)
+        return nn.ModuleList(layers)
+
+    def forward(self, x: Tensor) -> Tensor:
+        # See note [TorchScript super()]
+        for layer in self.layers:
+            x = layer(x)
+
+        return x
+
+
+def _resnet18(**kwargs):
+    return ResNet(BasicBlock, [2, 2, 2, 2], **kwargs)
+
+
+def _resnet34(**kwargs):
+    return ResNet(BasicBlock, [3, 4, 6, 3], **kwargs)
+
+
+def _resnet50(**kwargs):
+    return ResNet(Bottleneck, [3, 4, 6, 3], **kwargs)
+
+
+def _get_resnet(
+    preset: str = "resnet18",
+    block_type: str = "basic",
+    layers: List = [2, 2, 2, 2],
+    **kwargs,
+):
+    preset_models = {"resnet18": _resnet18, "resnet34": _resnet34, "resnet50": _resnet50}
+
+    if preset.lower() == "none":
+        block_dict = {"basic": BasicBlock, "bottleneck": Bottleneck}
+        block = block_dict[block_type]
+        return ResNet(block, layers, **kwargs)
+    elif preset in preset_models:
+        return preset_models[preset](**kwargs)
+    else:
+        raise KeyError(f"Specified preset {preset} is not implemented yet.")
+
+
+class ResNet_Encoder(nn.Module):
+    def __init__(
+        self,
+        preset: str = "resnet18",
+        block_type: str = "basic",
+        layers: List = [2, 2, 2, 2],
+        **kwargs,
+    ):
+        super().__init__()
+        self.network = _get_resnet(preset, block_type, **kwargs)
+        self.features = self.dim = self.network.features
+
+    def forward(self, x):
+        return self.network(x)
