+++ byol/utilities.py
@@ -19,27 +19,6 @@ paths = Path_Handler()
 path_dict = paths._dict()
 
 
-def entropy(p, eps=0.0000001, loss=False):
-    """
-    Calculate the entropy of a binary classification prediction given a probability for either of the two classes.
-
-    Keyword arguments:
-    eps -- small additive factor to avoid log(0)
-    loss -- boolean value determines whether to return detached value for inference (False) or differentiable value for training (True)
-    """
-    H_i = -torch.log(p + eps) * p
-    H = torch.sum(H_i, 1).view(-1)
-
-    if not loss:
-        # Clamp to avoid negative values due to eps
-        H = torch.clamp(H, min=0)
-        return H.detach().cpu().numpy()
-
-    H = torch.mean(H)
-
-    return H
-
-
 def fig2img(fig):
     """Convert a Matplotlib figure to a PIL Image and return it"""
     import io
@@ -64,27 +43,11 @@ def batch_eval(fn_dict, dset, batch_size=200):
         outs[key] = []
 
     for x, y in loader:
-        # Take only weakly augmented sample if passing through unlabelled data
-        #        if strong_T:
-        #            x = x[0]
-
         # Append result from each batch to list in outputs dictionary
         for key, fn in fn_dict.items():
             outs[key].append(fn(x, y))
 
     return outs
-
-
-def freeze_model(model):
-    for param in model.parameters():
-        param.requires_grad = False
-    # model.eval()
-
-
-def unfreeze_model(model):
-    for param in model.parameters():
-        param.requires_grad = True
-    # model.train()
 
 
 def _scheduler(
@@ -197,134 +160,111 @@ def log_examples(wandb_logger, dset, n=18):
     wandb_logger.log_image(key="image_pairs", images=save_list)
 
 
-def compute_encoded_mu_sig(dset, pl_module, batch_size=250):
-    """
-    Compute the mean and standard deviation of the encoded features of the dataset
-    """
-    # Load samples in batches
-    n_dset = len(dset)
-    loader = DataLoader(dset, batch_size)
-    channels = pl_module.config["model"]["features"]
+def rgz_cut(rgz_dset, threshold, mb_cut: bool = True, remove_duplicates=False):
+    """Cut rgz data-set based on angular size and whether data-point is contained in MiraBest"""
 
-    # Calculate mean
-    mu = 0
-    # print("Computing mean")
-    for x, _ in loader:
-        x = pl_module.encoder(x.to(pl_module.device)).squeeze()
-        weight = x.shape[0] / n_dset
-        mu += weight * torch.mean(x)
+    n = len(rgz_dset)
+    idx_bool = np.ones(n, dtype=bool)
+    idx = np.arange(n)
 
-    # Calculate std
-    D_sq = 0
-    # print("Computing std")
-    for x, _ in loader:
-        x = pl_module.encoder(x.to(pl_module.device)).squeeze()
-        D_sq += torch.sum(x - mu) ** 2
-    std = (D_sq / (n_dset * channels)) ** 0.5
+    if remove_duplicates:
+        idx_bool = np.zeros(n, dtype=bool)
+        _, idx_unique = np.unique(rgz_dset.data, axis=0, return_index=True)
+        idx_bool[idx_unique] = True
 
-    # print(f"mean: {mu}, std: {std}")
-    return mu, std.item()
+        print(f"Removed {n - np.count_nonzero(idx_bool)} duplicate samples")
+        n = np.count_nonzero(idx_bool)
 
+    idx_bool *= rgz_dset.sizes > threshold
+    print(f"Removing {n - np.count_nonzero(idx_bool)} samples below angular size threshold.")
+    n = np.count_nonzero(idx_bool)
 
-def check_unique_list(lst):
-    return len(set(lst)) == len(lst)
+    if mb_cut:
+        idx_bool *= rgz_dset.mbflg == 0
 
+        # Print number of MB samples removed
+        print(f"Removed {n - np.count_nonzero(idx_bool)} MiraBest samples from RGZ")
 
-"""
-Directly copied from pl_bolts
-https://github.com/PyTorchLightning/lightning-bolts/blob/0.3.0/pl_bolts/optimizers/lars_scheduling.py#L62-L81
+    idx = np.argwhere(idx_bool)
 
-References:
-    - https://github.com/NVIDIA/apex/blob/master/apex/parallel/LARC.py
-    - https://arxiv.org/pdf/1708.03888.pdf
-    - https://github.com/noahgolmant/pytorch-lars/blob/master/lars.py
-"""
+    subset = D.Subset(rgz_dset, idx)
+    print(f"RGZ dataset cut from {n} to {len(subset)} samples")
+    return subset
 
 
-class LARSWrapper(object):
-    """
-    Wrapper that adds LARS scheduling to any optimizer. This helps stability with huge batch sizes.
-    """
+def dset2tens(dset):
+    """Return a tuple (x, y) containing the entire input dataset (careful with large datasets)"""
+    return next(iter(DataLoader(dset, int(len(dset)))))
 
-    def __init__(self, optimizer, eta=0.02, clip=True, eps=1e-8):
-        """
-        Args:
-            optimizer: torch optimizer
-            eta: LARS coefficient (trust)
-            clip: True to clip LR
-            eps: adaptive_lr stability coefficient
-        """
-        self.optim = optimizer
-        self.eta = eta
-        self.eps = eps
-        self.clip = clip
-        # transfer optim methods
-        self.state_dict = self.optim.state_dict
-        self.load_state_dict = self.optim.load_state_dict
-        self.zero_grad = self.optim.zero_grad
-        self.add_param_group = self.optim.add_param_group
-        self.__setstate__ = self.optim.__setstate__
-        self.__getstate__ = self.optim.__getstate__
-        self.__repr__ = self.optim.__repr__
 
-    @property
-    def defaults(self):
-        return self.optim.defaults
+def compute_mu_sig_images(dset, batch_size=0):
+    """Compute mean and standard variance of a dataset (use batching with large datasets)"""
+    if batch_size:
+        # Load samples in batches
+        n_dset = len(dset)
+        loader = DataLoader(dset, batch_size)
+        n_channels = next(iter(loader))[0].shape[1]
 
-    @defaults.setter
-    def defaults(self, defaults):
-        self.optim.defaults = defaults
+        # Calculate mean
+        mu = torch.zeros(n_channels)
+        for x, _ in loader:
+            # print(x.shape, _.shape)
+            for c in np.arange(n_channels):
+                x_c = x[:, c, :, :]
+                weight = x.shape[0] / n_dset
+                mu[c] += weight * torch.mean(x_c).item()
 
-    @property
-    def __class__(self):
-        return Optimizer
+        # Calculate std
+        D_sq = torch.zeros(n_channels)
+        for x, _ in loader:
+            for c in np.arange(n_channels):
+                x_c = x[:, c, :, :]
+                D_sq += torch.sum((x_c - mu[c]) ** 2)
+        sig = (D_sq / (n_dset * x.shape[-1] * x.shape[-2])) ** 0.5
 
-    @property
-    def state(self):
-        return self.optim.state
+        mu, sig = tuple(mu.tolist()), tuple(sig.tolist())
+        print(f"mu: {mu}, std: {sig}")
+        return mu, sig
 
-    @property
-    def param_groups(self):
-        return self.optim.param_groups
+    else:
+        x, _ = dset2tens(dset)
+        n_channels = x.shape[1]
+        mu, sig = [], []
+        for c in np.arange(n_channels):
+            x_c = x[:, c, :, :]
+            mu.append(torch.mean(x).item())
+            sig.append(torch.std(x).item())
+        return tuple(mu), tuple(sig)
 
-    @param_groups.setter
-    def param_groups(self, value):
-        self.optim.param_groups = value
 
-    @torch.no_grad()
-    def step(self, closure=None):
-        weight_decays = []
+def compute_mu_sig_features(dset, batch_size=0):
+    """Compute mean and standard variance of a dataset (use batching with large datasets)"""
+    print("Computing mean and std of dataset")
+    if batch_size:
+        # Load samples in batches
+        n_dset = len(dset)
+        loader = DataLoader(dset, batch_size)
+        x, _ = next(iter(loader))
+        n, c, h, w = x.shape
 
-        for group in self.optim.param_groups:
-            weight_decay = group.get("weight_decay", 0)
-            weight_decays.append(weight_decay)
+        # Calculate mean
+        mu = 0
+        print("Computing mean")
+        for x, _ in tqdm(loader):
+            x = x
+            weight = x.shape[0] / n_dset
+            mu += weight * torch.mean(x)
 
-            # reset weight decay
-            group["weight_decay"] = 0
+        # Calculate std
+        D_sq = 0
+        print("Computing std")
+        for x, _ in tqdm(loader):
+            D_sq += torch.sum((x - mu) ** 2)
+        std = (D_sq / (n_dset * h * w)) ** 0.5
 
-            # update the parameters
-            [self.update_p(p, group, weight_decay) for p in group["params"] if p.grad is not None]
+        print(f"mean: {mu}, std: {std}")
+        return mu, std.item()
 
-        # update the optimizer
-        self.optim.step(closure=closure)
-
-        # return weight decay control to optimizer
-        for group_idx, group in enumerate(self.optim.param_groups):
-            group["weight_decay"] = weight_decays[group_idx]
-
-    def update_p(self, p, group, weight_decay):
-        # calculate new norms
-        p_norm = torch.norm(p.data)
-        g_norm = torch.norm(p.grad.data)
-
-        if p_norm != 0 and g_norm != 0:
-            # calculate new lr
-            new_lr = (self.eta * p_norm) / (g_norm + p_norm * weight_decay + self.eps)
-
-            # clip lr
-            if self.clip:
-                new_lr = min(new_lr / group["lr"], 1)
-
-            # update params with clipped lr
-            p.grad.data += weight_decay * p.data
-            p.grad.data *= new_lr
+    else:
+        x, _ = dset2tens(dset)
+        return torch.mean(x).item(), torch.std(x).item()
