--- config/byol/gzmnist.yml
@@ -1,60 +1,0 @@
-preset: null
-
-persistent_workers: False
-precision: 16  # [16, 32]-bit. 16 is auto. mixed precision not pure 16.
-
-model:
-    architecture: 'resnet18' 
-    features: 512     # Number of encoded features (null just uses original resnet output)
-    n_epochs: 400
-    downscale: True
-
-projection_head:
-    hidden: 1024
-    out: 256
-
-data:
-    color_channels: 3
-    classes: 4
-    input_height: 64
-    rotate: True
-    val_batch_size: 500
-
-linear:
-    n_epochs: 100
-    opt: 'sgd'
-    lr: 0.2
-    batch_size: 25
-    momentum: 0.9
-    weight_decay: 0
-    scheduler: 'cosine' # ['none', 'cosine'] ####
-
-knn:
-  neighbors: 20
-  temperature: 0.1
-
-## AUGMENTATIONS ##
-center_crop_size: null
-random_crop_scale: !!python/tuple [0.3, 1]
-s: 0.5
-p_grayscale: 0.3
-p_blur: 0.5 
-
-## Optimiser ## 
-# Base settings
-lr: 0.06
-momentum: 0.9
-weight_decay: 0.0005
-batch_size: 256
-opt: 'sgd'
-# Scheduler
-scheduler: 'cosine' # ['none', 'cosine', 'warmupcosine']
-warmup_epochs: 10  # Only used in linear warmup cosine annealing
-
-# LARS
-lars: False
-trust_coef: 0.003
-
-## BYOL ## 
-m: 0.99
-m_decay: False
