--- config/byol/imagenette-deepmind.yml
@@ -1,45 +1,0 @@
-preset: 'deepmind'
-
-model:
-    architecture: 'resnet18' 
-    features: 512     # Number of encoded features (null just uses original resnet output)
-    n_epochs: 800
-    downscale: False
-    m: 0.996 # Controls EMA hyperparameter for BYOL momentum encoder
-    m_decay: True
-
-projection_head:
-    hidden: 4096
-    out: 256
-
-data:
-    color_channels: 3
-    classes: 10
-    rotate: False
-    input_height: 128
-    batch_size: 256
-
-optimizer:
-  type: 'sgd'
-  lr: 0.2
-  momentum: 0.9
-  weight_decay: 0.000001
-  lars: False
-  trust_coef: 0.003
-
-scheduler:
-  type: 'warmupcosine' # ['none', 'cosine', 'warmupcosine']
-  warmup_epochs: 10  # Only used in linear warmup cosine annealing, if scheduler type is 'warmupcosine'
-
-augmentations:
-  s: 0.5
-  center_crop_size: 128
-
-linear:
-    n_epochs: 100
-    opt: 'sgd'
-    lr: 0.2
-    batch_size: 100
-    momentum: 0.9
-    nesterov: True
-    weight_decay: 0
