--- config/byol/imagenette.yml
@@ -1,45 +1,0 @@
-preset: 'deepmind'
-
-model:
-    architecture: 'resnet18' 
-    features: 512     # Number of encoded features (null just uses original resnet output)
-    n_epochs: 800 
-    downscale: False
-    m: 0.996 # Controls EMA hyperparameter for BYOL momentum encoder
-    m_decay: False
-
-projection_head:
-    hidden: 1024
-    out: 256
-
-data:
-    color_channels: 3
-    classes: 10
-    rotate: False
-    input_height: 128
-    batch_size: 256
-
-optimizer:
-  type: 'sgd'
-  lr: 0.06
-  momentum: 0.9
-  weight_decay: 0.0005
-  lars: False
-  trust_coef: 0.003
-
-scheduler:
-  type: 'cosine' # ['none', 'cosine', 'warmupcosine']
-  warmup_epochs: 10  # Only used in linear warmup cosine annealing, if scheduler type is 'warmupcosine'
-
-augmentations:
-  s: 0.5
-  center_crop_size: 128
-
-linear:
-    n_epochs: 150
-    opt: 'adam'
-    lr: 0.0005
-    batch_size: 512
-    momentum: 0
-    nesterov: False
-    weight_decay: 0
