--- config/byol/legs.yml
@@ -1,68 +1,0 @@
-preset: null
-
-# overriding global
-num_workers: 16
-prefetch_factor: 5
-persistent_workers: False
-precision: 16  # [16, 32]-bit. 16 is auto. mixed precision not pure 16. 
-
-model:
-    architecture: 'resnet18' 
-    features: 512     # Number of encoded features (null just uses original resnet output)
-    n_epochs: 400
-    downscale: False
-
-projection_head:
-    hidden: 1024
-    out: 256
-
-supervised_head:
-    hidden: 1024
-
-data:
-    color_channels: 3
-    input_height: 128  # TODO still maybe a bit low, but 224 doesn't fit well
-    classes: 2
-    rotate: True
-    val_batch_size: 512
-    precrop_size_ratio: 1.3
-
-linear:
-    n_epochs: 100
-    opt: 'sgd'
-    lr: 0.2
-    batch_size: 25
-    momentum: 0.9
-    weight_decay: 0
-    scheduler: 'cosine' # ['none', 'cosine'] ####
-
-knn:
-  neighbors: 100
-  temperature: 0.1
-
-## AUGMENTATIONS ##
-center_crop_size: null
-random_crop_scale: !!python/tuple [0.3, 0.85]
-random_crop_ratio: !!python/tuple [0.8, 1.2]
-s: 0.5  # color jitter strength
-p_grayscale: 0.3  # TODO consider
-p_blur: 0.
-
-## Optimiser ## 
-# Base settings
-lr: 0.06
-momentum: 0.9
-weight_decay: 0.0005
-batch_size: 512
-opt: 'adam'
-# Scheduler
-scheduler: 'none' # ['none', 'cosine', 'warmupcosine']
-warmup_epochs: 10  # Only used in linear warmup cosine annealing
-
-# LARS
-lars: False
-trust_coef: 0.003
-
-## BYOL ## 
-m: 0.99
-m_decay: False
