--- config/byol/stl10.yml
@@ -1,46 +1,0 @@
-preset: 'none'
-
-model:
-    architecture: 'resnet18' 
-    features: 512     # Number of encoded features (null just uses original resnet output)
-    n_epochs: 100 
-    downscale: False
-    m: 0.996 # Controls EMA hyperparameter for BYOL momentum encoder
-    m_decay: False
-
-projection_head:
-    hidden: 512
-    out: 128
-
-data:
-    color_channels: 3
-    classes: 10
-    rotate: False
-    input_height: 96
-    batch_size: 64
-
-optimizer:
-  type: 'sgd'
-  lr: 0.03
-  momentum: 0.9
-  weight_decay: 0.0004 # 0.0005
-  lars: False
-  trust_coef: 0.003
-
-scheduler:
-  type: 'cosine' # ['none', 'cosine', 'warmupcosine']
-  warmup_epochs: 10  # Only used in linear warmup cosine annealing, if scheduler type is 'warmupcosine'
-
-augmentations:
-  s: 1
-  center_crop_size: null
-
-linear:
-    n_epochs: 30
-    opt: 'adam'
-    lr: 0.0003
-    batch_size: 300
-    momentum: 0
-    nesterov: False
-    weight_decay: 0
-
