--- config/mae/rgz.yml
@@ -1,60 +1,0 @@
-preset: 'none'
-
-model:
-    n_epochs: 2
-    masking_ratio: 0.75
-
-    decoder: 
-        dim: 512
-        depth: 2
-        heads: 4
-        dim_head: 64
-
-    vit:
-        patch_size: 4
-        dim: 256
-        depth: 6
-        heads: 8
-        mlp_dim: 1028
-        # arch: 'vit-small'
-
-data:
-    color_channels: 1
-    classes: 2
-    rotate: True
-    input_height: 128 # Keep this at 128 and adjust center_crop_size
-    batch_size: 256
-    val_batch_size: 500
-    cut_threshold: 27
-    test_size: 0.3 
-
-optimizer:
-  type: 'adamw'
-  lr: 0.00015
-  momentum: 0.9
-  weight_decay: 0.0005 # 0.0005
-  lars: False
-  trust_coef: 0.003
-
-scheduler:
-  type: 'cosine' # ['none', 'cosine', 'warmupcosine']
-  warmup_epochs: 10  # Only used in linear warmup cosine annealing, if scheduler type is 'warmupcosine'
-
-linear:
-    n_epochs: 50
-    opt: 'adam'
-    lr: 0.0003
-    batch_size: 300
-    momentum: 0
-    nesterov: False
-    weight_decay: 0.5
-
-## DATA ## 
-augmentations:
-  center_crop_size: 64
-  random_crop_scale: !!python/tuple [0.5, 1]
-  p_radio: 0.5
-  p_blur: 0
-  blur_kernel: 3 # Odd non-integer number
-  s: 0.5
-  p_aug: 0.5
