--- config/supervised/rgz.yml
@@ -1,31 +1,0 @@
-preset: null
-
-model:
-    architecture: 
-      preset: 'resnet18' 
-      features: 512     # Number of encoded features (null just uses original resnet output)
-      n_epochs: 200
-      downscale: True
-
-data:
-    color_channels: 1
-    classes: 2
-    rotate: True
-    input_height: 128 # Keep this at 128 and adjust center_crop_size
-    batch_size: 50 
-    val_batch_size: 500
-    cut_threshold: 27
-    test_size: 0.2 
-    center_crop_size: 80
-
-    optimizer:
-      type: 'adam'
-      lr: 0.2 # Note this will be scaled with batch_size so best to leave
-      momentum: 0.9
-      weight_decay: 0.0000015 # 0.0005
-      beta_1: 0.9
-      beta_2: 0.999
-
-    scheduler:
-      decay_type: 'warmupcosine' # ['none', 'cosine', 'warmupcosine']
-      warmup_epochs: 10  # Only used in linear warmup cosine annealing, if scheduler type is 'warmupcosine'
